Split known state from common subexpression eliminator.
